removed dashaccount. created accounthandler.
