+++ src/dashboard/DashNavBar.js
@@ -7,8 +7,7 @@ import DashData from "./DashData";
 import Tools from "./Tools/Tools";
 import UsersQuery from "./Tools/UsersQuery";
 import DashAbout from "./DashAbout";
-import DashAccount from "./DashAccount";
-import DashAccountFree from "./DashAccountFree";
+import AccountHandler from "./AccountHandler";
 import DashLoginModal from "./DashLoginModal";
 import DashSignupModal from "./DashSignupModal";
 import DashLogout from "./DashLogout";
@@ -74,7 +73,7 @@ function DashNav() {
       <Route exact path="/login" component={Login} />
       <Route exact path="/signup" component={Signup} />
       {/* <Route exact path="/services" component={DashAbout} /> */}
-      <ProtectedRoute exact path="/myaccount" component={DashAccountFree} />
+      <ProtectedRoute exact path="/myaccount" component={AccountHandler} />
       <ProtectedRoute exact path="/logout" component={DashLogout} />
     </>
   );
